% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargZIfit.R
\name{margZIfit.formula}
\alias{margZIfit.formula}
\title{Formula interface for fitting marginal ZI regression models}
\usage{
\method{margZIfit}{formula}(
  formula,
  data = list(),
  idvar,
  timevar,
  method,
  mulink,
  plink = "logit",
  errDist,
  corstruct = "indep",
  mixcorstr = "indep",
  nophi
)
}
\arguments{
\item{formula}{Two part formula, first part non-degenerate, second part degenerate  e.g. y ~ 1 + a | 1}

\item{data}{Specify data}

\item{idvar}{Specify the cluster specific ID variable}

\item{timevar}{Specify the time or measurement occassion variable}

\item{method}{Method for fitting: one of ES_BinMix (Es binary mixing), ES_MultiMix (ES multinomial mixing), RS_MultiMix (RS multinomial mixing)}

\item{mulink}{Link function for the non-degenerate component ("log","sqrt","logit","probit","identity","cloglog","inverse")}

\item{plink}{Link function for the degenerate component ("logit","probit")}

\item{errDist}{Distribution for the non-degenerate component ("poisson","binomial")}

\item{corstruct}{Correlation structure for the non-degenerate component ("indep","exch","ar1","unspec","onedep","toep3"), default = "indep"}

\item{mixcorstr}{Correlation structure for the degenerate component ("indep","exch","ar1","unspec","onedep","toep3"), default = "indep"}

\item{nophi}{Logical: Whether dispersion parameter is to be estimated or not, default = FALSE}
}
\value{
model parameters and standard errors
}
\description{
Formula interface for fitting marginal ZI regression models
}
