% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargZIfit.R
\name{margZIfitEst}
\alias{margZIfitEst}
\title{Fit marginal ZI regression model passed from a formula interface or directly using the model parameters}
\usage{
margZIfitEst(
  bmat,
  gmat,
  y,
  params,
  method,
  mulink,
  plink = "logit",
  formula2,
  dims,
  id,
  time,
  nClust,
  clustSize,
  bin,
  errDist,
  corstruct = "indep",
  mixcorstr = "indep",
  nophi,
  verbose = FALSE
)
}
\arguments{
\item{bmat}{Design matrix for the non-degenerate component}

\item{gmat}{Design matrix for the degenerate component}

\item{y}{Response variable}

\item{params}{Vector of intial parametrs to the model}

\item{method}{Method for fitting: one of ES_BinMix (Es binary mixing), ES_MultiMix (ES multinomial mixing), RS_MultiMix (RS multinomial mixing)}

\item{mulink}{Link function for the non-degenerate component ("log","sqrt","logit","probit","identity","cloglog","inverse")}

\item{plink}{Link function for the degenerate component ("logit","probit")}

\item{formula2}{Formula object for the degenerate component}

\item{dims}{Vector or list of parameter dimensions by component}

\item{id}{Cluster fic ID variable}

\item{time}{Time or measurement occassion variable}

\item{nClust}{Number of clusters (sample size)}

\item{clustSize}{Number of observations in a cluster}

\item{bin}{Vector of binomial denominators}

\item{errDist}{Distribution for the non-degenerate component ("poisson","binomial")}

\item{corstruct}{Correlation structure for the non-degenerate component ("indep","exch","ar1","unspec","onedep","toep3"), default = "indep"}

\item{mixcorstr}{Correlation structure for the degenerate component ("indep","exch","ar1","unspec","onedep","toep3"), default = "indep"}

\item{nophi}{Logical: Whether dispersion parameter is to be estimated or not, default = FALSE}

\item{verbose}{Logical: Wheteher to print out model fitting steps, default = FALSE}
}
\value{
list of model parameters and standard errors.
}
\description{
Fit marginal ZI regression model passed from a formula interface or directly using the model parameters
}
