% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nHosp}
\alias{nHosp}
\title{Number of hospital visits data (Karim and Zeger, 1988)}
\format{A data frame with 73 rows and 8 variables:
\describe{
  \item{V1}{child's ID}
  \item{V2}{child's age in months}
  \item{V3}{child's gender}
  \item{V4}{smoking status}
  \item{V5}{Counts for quarter 1}
  \item{V6}{Counts for quarter 2}
  \item{V7}{Counts for quarter 3}
  \item{V8}{Counts for quarter 4}
}}
\source{
Karim and Zeger, 1988
}
\usage{
nHosp
}
\description{
A data in wide format that describes the number of hospital visits by children who were followed through the foru quarters of the year.
}
\keyword{datasets}
